Attachments can now attach anywhere on a rail
Added the ability for an attachment to be able to be placed anywhere on a rail, instead of just in the center. A held attachment will snap to the rail if close enough and it's position will be based on the closest point on the rail to the controller the attachment is held by.